Move 'WeakRefProxyInstance' and its compositions to a separate file
