+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeediOS/Feed UI/Composer/Helpers/WeakRefProxyInstance.swift
@@ -1,0 +1,28 @@
+//
+//  WeakRefProxyInstance.swift
+//  EssentialFeediOS
+//
+//  Created by Fardan Akhter on 11/05/2023.
+//
+
+import UIKit
+
+final class WeakRefProxyInstance<T: AnyObject>{
+    private(set) weak var instance: T?
+    
+    init(_ instance: T?) {
+        self.instance = instance
+    }
+}
+
+extension WeakRefProxyInstance: FeedLoadingView where T: FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel) {
+        instance?.display(viewModel)
+    }
+}
+
+extension WeakRefProxyInstance: FeedImageView where T: FeedImageCellController {
+    func display(_ viewModel: FeedImageViewModel<UIImage>) {
+        instance?.display(viewModel)
+    }
+}
